nit: cleanup
