+++ packages/core/src/tests/fixtures.ts
@@ -1,27 +1,7 @@
 /**
  * Nodes
  */
-
-export const createTestNode = (id, data, config = {}) => {
-  return {
-    id,
-    data: {
-      props: {},
-      custom: {},
-      hidden: false,
-      isCanvas: false,
-      ...data,
-    },
-    related: {},
-    events: { selected: false, dragged: false, hovered: false },
-    rules: {
-      canMoveIn: () => true,
-      canMoveOut: () => true,
-      canDrag: () => true,
-    },
-    ...config,
-  };
-};
+import { createTestNode } from "../utils/createTestNode";
 
 export const rootNode = createTestNode("ROOT", {
   name: "Document",
