--- packages/utils/nodemon.json
@@ -1,3 +1,0 @@
-{
-  "ext": "js,jsx,json,ts,tsx"
-}
